refactored types and Components to separate modules to benefit dev environment
