+++ src/App.tsx
@@ -1,133 +1,9 @@
 import { useState, useEffect } from "react";
 import "./App.css";
-import FileUploader from "./FileUploader";
+import FileUploader from "./Components/FileUploader.tsx";
 import { lazyToggleTheme } from "./utilities.ts";
-
-export interface OptimizationSettings {
-  enableDedup: boolean;
-  dedupOptions: {
-    accessors: boolean;
-    meshes: boolean;
-    materials: boolean;
-  };
-  enableTextureCompression: boolean;
-  textureCompressionOptions: {
-    format: "webp" | "jpeg" | "png";
-    quality: "auto" | number;
-    resize: [number, number] | null;
-  };
-  enableFlatten: boolean;
-  enableJoin: boolean;
-  enableWeld: boolean;
-  enableSimplify: boolean;
-  simplifyOptions: {
-    ratio: number;
-    error: number;
-  };
-  enableCenter: boolean;
-  centerOptions: {
-    pivot: "center" | "bottom" | "origin";
-  };
-  enableMeshopt: boolean;
-  meshoptOptions: {
-    level: "high" | "medium" | "low";
-  };
-  enablePrune: boolean;
-  pruneOptions: {
-    keepExtras: boolean;
-  };
-  enableQuantize: boolean;
-  enableResample: boolean;
-  enableInstance: boolean;
-  instanceOptions: {
-    min: number;
-  };
-  enableSparse: boolean;
-  sparseOptions: {
-    ratio: number;
-  };
-  enablePalette: boolean;
-  paletteOptions: {
-    min: number;
-  };
-  enableNormals: boolean;
-  normalsOptions: {
-    overwrite: boolean;
-  };
-  enableMetalRough: boolean;
-  enableMaterialsOptions: boolean;
-  materialsOptions: {
-    doubleSided: boolean;
-  };
-  userSettings: {
-    fileNameSuffix: string;
-    maxFileNameLength: number;
-    shortenFileNames: boolean;
-  };
-}
-
-const defaultSettings: OptimizationSettings = {
-  enableDedup: true,
-  dedupOptions: {
-    accessors: true,
-    meshes: true,
-    materials: true,
-  },
-  enableTextureCompression: true,
-  textureCompressionOptions: {
-    format: "webp",
-    quality: "auto",
-    resize: [1024, 1024],
-  },
-  enableFlatten: true,
-  enableJoin: true,
-  enableWeld: true,
-  enableSimplify: false, // Changed from true to false
-  simplifyOptions: {
-    ratio: 0.75,
-    error: 0.01,
-  },
-  enableCenter: false,
-  centerOptions: {
-    pivot: "center",
-  },
-  enableMeshopt: false,
-  meshoptOptions: {
-    level: "medium",
-  },
-  enablePrune: true,
-  pruneOptions: {
-    keepExtras: true,
-  },
-  enableQuantize: false,
-  enableResample: false,
-  enableInstance: false,
-  instanceOptions: {
-    min: 5,
-  },
-  enableSparse: false,
-  sparseOptions: {
-    ratio: 0.1,
-  },
-  enablePalette: false,
-  paletteOptions: {
-    min: 3,
-  },
-  enableNormals: false,
-  normalsOptions: {
-    overwrite: true,
-  },
-  enableMetalRough: false,
-  enableMaterialsOptions: false,
-  materialsOptions: {
-    doubleSided: false,
-  },
-  userSettings: {
-    fileNameSuffix: "_optimized",
-    maxFileNameLength: 20,
-    shortenFileNames: false,
-  },
-};
+import { defaultSettings } from "./settings.ts";
+import type { OptimizationSettings } from "./types.ts";
 
 function App() {
   const [settings, setSettings] = useState<OptimizationSettings>(() => {
